financial-statement-summarizer/requirements.txt
openai==0.28.0
python-dotenv==1.0.1
streamlit==1.37.1
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-text-splitters==0.2.2
langsmith==0.1.99